Refactor layout and page structure; remove subscription page and update user profile dropdown
- Updated RootLayout to wrap children in a main element with max width and padding.
- Refactored HomePage to remove redundant main element and streamline card rendering logic.
- Deleted the subscription page and its associated logic, including checkout and customer portal functionalities.
- Modified UserProfileDropdown to replace Subscription & Billing link with Settings link and icon.
