+++ apps/web/app/layout.tsx
@@ -29,7 +29,7 @@ export default function RootLayout({
               disableTransitionOnChange
               enableSystem={true}
             >
-              {children}
+              <main className="max-w-7xl mx-auto px-4 pb-10">{children}</main>
             </ThemeProvider>
           </ConvexQueryCacheProvider>
         </ConvexClientProvider>
